Add helper for user registration and integrate across events
